+++ source/ChanSort.Loader.Sony/SerializerPlugin.cs
@@ -1,0 +1,16 @@
+﻿using ChanSort.Api;
+
+namespace ChanSort.Loader.Sony
+{
+  public class SerializerPlugin : ISerializerPlugin
+  {
+    public string DllName { get; set; }
+    public string PluginName => "Sony sdb.xml";
+    public string FileFilter => "sdb*.xml";
+
+    public SerializerBase CreateSerializer(string inputFile)
+    {
+      return new Serializer(inputFile);
+    }
+  }
+}